chore: Update models and traits
This commit updates the `KanbanScrumHelper` trait to replace the `code` attribute with `slug` in the `getRecords` method. It also adds the `slug` attribute to the `IssueType` model and the `IssueSubscriber` model. Additionally, it adds the `slug` attribute and relationships to the `IssueRelation` model. Finally, it updates the `Project` model by fixing a typo in the `icon` method and adding relationships to the `IssuePriority` model.