feat: Add ImportJiraIssues job for importing Jira issues
This commit adds a new job called `ImportJiraIssues` that is responsible for importing Jira issues into the application. The job takes an array of issues and a user as parameters. It iterates over each issue, retrieves the project details and issue data, and creates the corresponding project and issue records in the database. If the project does not exist, it creates a new project with the necessary details. It also creates a `ProjectUser` record to associate the user with the project. Finally, it creates an `Issue` record with the issue details. After importing the issues, a notification is sent to the user to indicate the successful importation.

```