feat: Add IssueCommented notification
This commit adds a new notification class called `IssueCommented`. This notification is triggered when a new comment is added to an issue. It sends an email notification to the relevant users and also stores the notification in the database. The email notification includes the details of the issue and a link to view the issue. This notification improves the communication and collaboration among users in the application.