feat: Add ViewAction to Filament resource pages
This commit adds the ViewAction to the Filament resource pages for EditIssue, EditActivity, EditCalendar, EditIssueStatus, and EditDesignElements. The ViewAction is included in the `getHeaderActions` method of each respective page class. This enhancement improves the user interface of the application by providing the ability to view the details of each resource directly from the resource page.