feat: Add IsPermissable trait for checking permissions
This commit adds the IsPermissable trait to the app's Traits directory. The IsPermissable trait serves as a placeholder for any methods that need to be added to models to check permissions, as well as a way to identify models that have permissions. Currently, it is a placeholder/dummy trait and needs to be updated with methods to check permissions.