feat: Add HasRoles trait for managing user roles and permissions
This commit adds the HasRoles trait to the app's Traits directory. The HasRoles trait is used to manage user roles and permissions. It includes the Spatie\Permission\Traits\HasRoles trait and the HasPermissions trait, allowing users to have multiple roles and permissions. This trait will be used to set and check permissions for the User model.