feat: Add user deletion functionality and password confirmation test
The code changes include adding user deletion functionality and a corresponding test for password confirmation. The `DeleteAccountTest` now tests that user accounts can be deleted by calling the `deleteUser` method in the `DeleteUserForm` Livewire component. It also checks if the user was successfully deleted or soft deleted by verifying the absence of the user in the database.

In the `PasswordConfirmationTest`, the test case `testPasswordIsNotConfirmedWithInvalidPassword` now asserts that the response is a redirect and that the session has errors for the 'password' field.