feat: Seed permission sets in DatabaseSeeder
This commit adds the code to seed the permission sets in the DatabaseSeeder class. The new code calls the PermissionSetSeeder class to create dynamic permission sets based on models and available permissions. This enhances the functionality of the application by providing a flexible and automated way to manage permissions for different models.