refactor: Add protected field to roles table
This commit adds a new migration file to add a protected field to the roles table. The protected field is a boolean column that defaults to false. This field is used to protect certain roles from being deleted for security reasons. The migration file also includes the necessary code to add the protected field to the roles table in the database. This change enhances the role management functionality by allowing certain roles to be protected and preventing accidental deletion.