refactor: Add Jira and Gitea settings classes
This commit adds the JiraSettings and GiteaSettings classes to the app/Settings directory. These classes extend the Spatie\LaravelSettings\Settings class and define the settings for the Jira and Gitea features respectively.

The JiraSettings class sets the group name to 'default' and does not have any additional properties or methods currently.

The GiteaSettings class sets the group name to 'gitea' and defines the following properties:
- enabled: a boolean indicating whether the Gitea module is enabled or not.
- base_url: a string representing the base URL of the Gitea server instance.
- api_token: a string representing the API token used to authenticate with the Gitea server instance.

These classes will be used to store and retrieve the settings for the Jira and Gitea features in the application.